Refactor enterprise code
