+++ worker/acl_cache.go
@@ -14,14 +14,10 @@ package worker
 
 import (
 	"encoding/json"
-	"fmt"
-	"strings"
 	"sync"
 
 	"github.com/dgraph-io/dgraph/ee/acl"
-	"github.com/dgraph-io/dgraph/x"
 	"github.com/golang/glog"
-	"github.com/pkg/errors"
 )
 
 // aclCache is the cache mapping group names to the corresponding group acls
@@ -78,29 +74,6 @@ func (cache *aclCache) Update(groups []acl.Group) {
 	AclCachePtr.predPerms = predPerms
 }
 
-func (cache *aclCache) authorizePredicate(groups []string, predicate string,
-	operation *acl.Operation) error {
-	if x.IsAclPredicate(predicate) {
-		return errors.Errorf("only groot is allowed to access the ACL predicate: %s", predicate)
-	}
-
-	AclCachePtr.RLock()
-	predPerms := AclCachePtr.predPerms
-	AclCachePtr.RUnlock()
-
-	if groupPerms, found := predPerms[predicate]; found {
-		if hasRequiredAccess(groupPerms, groups, operation) {
-			return nil
-		}
-	}
-
-	// no rule has been defined that can match the predicate
-	// by default we block operation
-	return errors.Errorf("unauthorized to do %s on predicate %s",
-		operation.Name, predicate)
-
-}
-
 // hasRequiredAccess checks if any group in the passed in groups is allowed to perform the operation
 // according to the acl rules stored in groupPerms
 func hasRequiredAccess(groupPerms map[string]int32, groups []string,
@@ -112,56 +85,4 @@ func hasRequiredAccess(groupPerms map[string]int32, groups []string,
 		}
 	}
 	return false
-}
-
-func AuthorizePreds(userId string, groupIds, preds []string,
-	aclOp *acl.Operation) map[string]struct{} {
-
-	blockedPreds := make(map[string]struct{})
-	for _, pred := range preds {
-		if err := AclCachePtr.authorizePredicate(groupIds, pred, aclOp); err != nil {
-			logAccess(&accessEntry{
-				userId:    userId,
-				groups:    groupIds,
-				preds:     preds,
-				operation: aclOp,
-				allowed:   false,
-			})
-
-			blockedPreds[pred] = struct{}{}
-		}
-	}
-	return blockedPreds
-}
-
-type accessEntry struct {
-	userId    string
-	groups    []string
-	preds     []string
-	operation *acl.Operation
-	allowed   bool
-}
-
-func NewAccessEntry(userId string, groups, preds []string,
-	operation *acl.Operation, allowed bool) *accessEntry {
-
-	return &accessEntry{
-		userId:    userId,
-		groups:    groups,
-		preds:     preds,
-		operation: operation,
-		allowed:   allowed,
-	}
-}
-
-func logAccess(log *accessEntry) {
-	if glog.V(1) {
-		glog.Info(log.String())
-	}
-}
-
-func (log *accessEntry) String() string {
-	return fmt.Sprintf("ACL-LOG Authorizing user %q with groups %q on predicates %q "+
-		"for %q, allowed:%v", log.userId, strings.Join(log.groups, ","),
-		strings.Join(log.preds, ","), log.operation.Name, log.allowed)
 }
